backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.8.0
pydantic==2.9.2
pydantic-settings>=2.5.2
python-dotenv==1.0.1
httpx>=0.27.2
tavily-python==0.5.0
# LangChain RAG stack
langchain>=1.2.0
langchain-community>=0.4.0
langchain-google-genai>=4.2.0
langchain-text-splitters>=1.1.0
faiss-cpu>=1.9.0

# Auth
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
